refactor: remove dead Telegram handlers, fix latency scan bug
Delete 609 lines of dead handler code — handleHelp, handleClear,
handleStats, handleForget, handleFacts, handleReflect, handleReflections,
handlePersona (+ traits/history/rewrite), handleDream, handleDreamLog,
handleLastTrace, handleCompact, handleStatus, handleTraces.

All replaced by Exec* methods in commands_gateway.go which serve
both Telegram and Gradio through the unified command system.

Fix: AVG(latency_ms) returns float64 from SQLite but was scanned
into int, causing a warning on every /stats call. Added CAST AS
INTEGER in the SQL query.
